update: use java.util.BitSet on js platform
scala.js supports java.util.BitSet since 1.9.0
https://www.scala-js.org/news/2022/02/14/announcing-scalajs-1.9.0/
